commited by Abner

+++ ze-delivery-spring-boot-back-end-java/src/main/java/com/org/ze_delivery_spring_boot_back_end_java/models/services/interfaces/IPartnersService.java
@@ -1,5 +1,7 @@
 package com.org.ze_delivery_spring_boot_back_end_java.models.services.interfaces;
 
+import java.math.BigDecimal;
+
 import com.org.ze_delivery_spring_boot_back_end_java.models.dtos.requests.PartnerRequest;
 import com.org.ze_delivery_spring_boot_back_end_java.models.dtos.responses.PartnerResponse;
 
@@ -8,5 +10,5 @@ public interface IPartnersService {
 	
 	PartnerResponse loadPartnerById(String id);
 	
-	PartnerResponse searchPartnerByLongAndLat(Long longitude, Long latitude);
+	PartnerResponse searchPartnerByLongAndLat(BigDecimal longitude, BigDecimal latitude);
 }
